Add new setting to keep outdated PR comments
Default for `sonar.github.deleteOutdatedInlineComments` is `true` to keep backwards compatibility.

We prefer to set it `false` and contribute this PR.

Thanks for this great tool :+1:

+++ src/main/java/org/sonar/plugins/github/PullRequestIssuePostJob.java
@@ -60,7 +60,9 @@ public class PullRequestIssuePostJob implements PostJob {
 
     updateReviewComments(commentsToBeAddedByLine);
 
-    pullRequestFacade.deleteOutdatedComments();
+    if (gitHubPluginConfiguration.shouldDeleteOutdatedInlineComments()) {
+      pullRequestFacade.deleteOutdatedComments();
+    }
 
     pullRequestFacade.createOrUpdateGlobalComments(report.hasNewIssue() ? report.formatForMarkdown() : null);
